refactor: move all render headers to render folder
