+++ include/render/hit_record.h
