+++ include/render/hittables/plane.h
@@ -1,8 +1,8 @@
 #ifndef PLANE_H
 #define PLANE_H
 
-#include "hittables/hittable.h"
-#include "ray.h"
+#include "render/hittables/hittable.h"
+#include "render/ray.h"
 
 t_hittable_data	plane_new(void);
 bool 			plane_hit(const t_hittable *hittable, const t_ray *ray, t_hit_record *hit_record);
