+++ src/render/render_params.c
@@ -1,4 +1,4 @@
-#include "render.h"
+#include "render/render.h"
 #include <stddef.h>
 #include <stdlib.h>
